+++ CHANGELOG.md
@@ -170,6 +170,9 @@
   highlighting whole line when line number and character positioning is used
 - [#218](https://github.com/estruyf/vscode-demo-time/issues/218): Added a new
   `/api/demos` endpoint to retrieve the list of demos in the workspace
+- [#330](https://github.com/estruyf/vscode-demo-time/issues/330): Added new `
+  copyFromSelection` action to copy text from the current selection in the
+  editor
 
 ## [1.7.1] - 2025-07-10